refactor(storage): extract authenticated construction shaping merges (#1184)
